Specify s3Bucket.quota as a "number" (instead of a "integer")
`quota` should be a u64 since it describes bytes sizes.

The Pulumi schema doesn't support u64 (or i64), so we are converting to a float64, which
should get us at least 2^52 bits of precision while minimizing the breaking change (int ->
float) for our users.

Fixes https://github.com/pulumi/pulumi-minio/issues/449

+++ sdk/dotnet/S3Bucket.cs
@@ -59,7 +59,7 @@ namespace Pulumi.Minio
         public Output<bool?> ObjectLocking { get; private set; } = null!;
 
         [Output("quota")]
-        public Output<int?> Quota { get; private set; } = null!;
+        public Output<double?> Quota { get; private set; } = null!;
 
 
         /// <summary>
@@ -123,7 +123,7 @@ namespace Pulumi.Minio
         public Input<bool>? ObjectLocking { get; set; }
 
         [Input("quota")]
-        public Input<int>? Quota { get; set; }
+        public Input<double>? Quota { get; set; }
 
         public S3BucketArgs()
         {
@@ -155,7 +155,7 @@ namespace Pulumi.Minio
         public Input<bool>? ObjectLocking { get; set; }
 
         [Input("quota")]
-        public Input<int>? Quota { get; set; }
+        public Input<double>? Quota { get; set; }
 
         public S3BucketState()
         {
