+++ sdk/java/src/main/java/com/pulumi/minio/S3BucketArgs.java
@@ -6,7 +6,7 @@ package com.pulumi.minio;
 import com.pulumi.core.Output;
 import com.pulumi.core.annotations.Import;
 import java.lang.Boolean;
-import java.lang.Integer;
+import java.lang.Double;
 import java.lang.String;
 import java.util.Objects;
 import java.util.Optional;
@@ -53,9 +53,9 @@ public final class S3BucketArgs extends com.pulumi.resources.ResourceArgs {
     }
 
     @Import(name="quota")
-    private @Nullable Output<Integer> quota;
+    private @Nullable Output<Double> quota;
 
-    public Optional<Output<Integer>> quota() {
+    public Optional<Output<Double>> quota() {
         return Optional.ofNullable(this.quota);
     }
 
@@ -133,12 +133,12 @@ public final class S3BucketArgs extends com.pulumi.resources.ResourceArgs {
             return objectLocking(Output.of(objectLocking));
         }
 
-        public Builder quota(@Nullable Output<Integer> quota) {
+        public Builder quota(@Nullable Output<Double> quota) {
             $.quota = quota;
             return this;
         }
 
-        public Builder quota(Integer quota) {
+        public Builder quota(Double quota) {
             return quota(Output.of(quota));
         }
